requirements.txt
Flask==2.2.2
Flask-Injector==0.14.0
greenlet==2.0.1
gunicorn==20.1.0
Flask-JWT-Extended==4.4.4